Initial code review solving issues: Mostly moving stuff around. Concentrated all synergy specific inside of ai/requested_scaling
Todo: still need to expand on the tests and build up a consistent tests folder
