+++ rs/ai/requested_scaling/handlers/synergy_handlers/exhaust_action/exhaust_test_fixture.py
