+++ rs/ai/requested_scaling/requested_scaling.py
@@ -21,7 +21,7 @@ from rs.ai.requested_scaling.handlers.shop_entrance_handler import ShopEntranceH
 from rs.ai.requested_scaling.handlers.shop_purchase_handler import ShopPurchaseHandler
 from rs.ai.requested_scaling.handlers.smart_battle_handler import SmartBattleHandler
 from rs.ai.requested_scaling.handlers.smart_path_handler import SmartPathHandler
-from rs.ai.requested_scaling.handlers.synergy_handlers.exhaust_handler import ExhaustHandler
+from rs.ai.requested_scaling.handlers.synergy_handlers.exhaust_action.exhaust_handler import ExhaustHandler
 from rs.ai.requested_scaling.handlers.synergy_handlers.scaling_battle_handler import ScalingBattleHandler
 from rs.ai.requested_scaling.handlers.synergy_handlers.synergy_card_reward_handler import SynergyCardRewardHandler
 from rs.ai.requested_scaling.handlers.transform_handler import TransformHandler
