+++ rs/ai/requested_scaling/synergies/synergy_tags.py
