+++ rs/ai/requested_scaling/test/fixtures/scaling_text_fixture.py
