+++ rs/ai/requested_scaling/test/fixtures/synergy_choosing_text_fixture.py
