+++ rs/ai/requested_scaling/test/synergy_states/handle_scaling_with_no_energy.json
