+++ rs/ai/requested_scaling/test/synergy_states/how_to_handle_exhausted_cards.json
