+++ rs/ai/requested_scaling/test/synergy_states/learn_to_choose_cards.json
