+++ rs/ai/requested_scaling/test/synergy_states/should_play_fnp.json
