+++ rs/ai/requested_scaling/test/synergy_states/should_skip_to_discard_handler.json
