+++ rs/ai/requested_scaling/test/synergy_states/smarter_card_reward_testing.json
