+++ rs/ai/requested_scaling/test/synergy_states/stuck_synergy.json
