+++ rs/ai/requested_scaling/test/test_scaling_battle_handler.py
