+++ rs/ai/requested_scaling/test/test_synergy_card_reward_handler.py
