+++ rs/game/current_action.py
@@ -2,5 +2,5 @@ from enum import Enum
 
 
 class CurrentAction(Enum):
-    EXHAUSTACTION = 'ExhaustAction'
+    EXHAUST_ACTION = 'ExhaustAction'
